Split parser funcs into relevant expresser files.
Also renamed IMExIterator to IMExpresser, and added parse method to it.
However, this parse method requires knowledge of T and I to call now, so it can't be used in the constructors,
I'll probably need to remove parse from the trait and just define the parse method in a normal impl.
Or alternatively, make a different trait for parsable expressers.
This solution is probably better since there's not much of a connection between the expressers being able to parse
and them being able to iterate.
Also, this way I can implement the parsable trait on Quantifier, which doesn't iterate.

All of these traits may be pointless though, since these functions probably won't be used generically anyway.
Still, it's a nice exercise and future-proofing in case they are used generally one day.