Introduced new trait ParserCombinator.
This splits the IMExpresser trait into ParserCombinator and IMExIterator again.
This allows having a trait with a parse fn that doesn't require knowledge of T, I to use.
However, the way these imports work, now all parse functions aren't usable outside the library.
This is probably fine, unless somebody wants to use them someday, they are welcome to make the case for their publicity.
As a result, I had to remove one doc example that tries to import a parser func.