Attempting to add an IMExIterCounter struct for assisting with state in quantified_imex_val.rs
Also experimenting with using associated types instead of generics on IMExIterator. Attempting to make IMExIterCounter with generics
was troublesome.
The reason being that creating an instance of IMExIterCounter requires knowledge of the types involved in the iterate function.
The entire tree of expression objects SHOULDN'T know what they're iterating over, by design, so this was unhelpful.
Hoping that associated types in the IMExIterator trait will help improve things so types involved in iteration can remain unkown until they really need to be known.